requirements.txt
khl.py
aiomysql
tortoise-orm
pydantic
colorlog~=6.7.0
colorama~=0.4.6
KaiHeiLaCardBuilder
psutil
APScheduler